Reimplementation from clean, without yo
