+++ .vscode/launch.json
@@ -32,7 +32,7 @@
 		},
 		{
 			"name": "Run Web Extension ",
-			"type": "pwa-extensionHost",
+			"type": "extensionHost",
 			"debugWebWorkerHost": true,
 			"request": "launch",
 			"args": [
