+++ .vscode/settings.json
@@ -7,5 +7,5 @@
         "out": true // set this to false to include "out" folder in search results
     },
     // Turn off tsc task auto detection since we have the necessary tasks as npm scripts
-    "typescript.tsc.autoDetect": "off"
+    "js/ts.tsc.autoDetect": "off"
 }
