+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 All notable changes to the "vscode-uuid-generator" extension will be documented in this file.
 
+## [1.1.0] - 2026-04-21
+Rewrite to eliminate unnecessary dependencies to reduce exposure to vulnerabilities
+
 ## [1.0.1] - 2026-02-11
 Updated dependencies
 
@@ -43,6 +46,7 @@ Initial release to go through the process of preparing and uploading
     - Copy UUID to clipboard
     - Insert UUID into current editor
 
+[1.1.0]: https://github.com/Motivesoft/vscode-uuid-generator/releases/tag/v1.1.0
 [1.0.1]: https://github.com/Motivesoft/vscode-uuid-generator/releases/tag/v1.0.1
 [1.0.0]: https://github.com/Motivesoft/vscode-uuid-generator/releases/tag/v1.0.0
 [0.4.1]: https://github.com/Motivesoft/vscode-uuid-generator/releases/tag/v0.4.1
